+++ features/step_definitions/matching.js
@@ -6,7 +6,8 @@ module.exports = function () {
     this.When(/^I match I should get$/, (table, callback) => {
         var got;
 
-        this.reprocessAndLoadData(() => {
+        this.reprocessAndLoadData((e) => {
+            if (e) return callback(e);
             var testRow = (row, ri, cb) => {
                 var afterRequest = (err, res) => {
                     if (err) return cb(err);