+++ include/extractor/guidance/turn_handler.hpp
@@ -5,6 +5,7 @@
 #include "extractor/guidance/intersection_handler.hpp"
 #include "extractor/query_node.hpp"
 
+#include "util/attributes.hpp"
 #include "util/name_table.hpp"
 #include "util/node_based_graph.hpp"
 
@@ -46,15 +47,19 @@ class TurnHandler : public IntersectionHandler
                         const ConnectedRoad &road,
                         const ConnectedRoad &other) const;
     // Dead end.
+    OSRM_ATTR_WARN_UNUSED
     Intersection handleOneWayTurn(Intersection intersection) const;
 
     // Mode Changes, new names...
+    OSRM_ATTR_WARN_UNUSED
     Intersection handleTwoWayTurn(const EdgeID via_edge, Intersection intersection) const;
 
     // Forks, T intersections and similar
+    OSRM_ATTR_WARN_UNUSED
     Intersection handleThreeWayTurn(const EdgeID via_edge, Intersection intersection) const;
 
     // Handling of turns larger then degree three
+    OSRM_ATTR_WARN_UNUSED
     Intersection handleComplexTurn(const EdgeID via_edge, Intersection intersection) const;
 
     void
@@ -64,9 +69,12 @@ class TurnHandler : public IntersectionHandler
     std::pair<std::size_t, std::size_t> findFork(const EdgeID via_edge,
                                                  const Intersection &intersection) const;
 
+    OSRM_ATTR_WARN_UNUSED
     Intersection assignLeftTurns(const EdgeID via_edge,
                                  Intersection intersection,
                                  const std::size_t starting_at) const;
+
+    OSRM_ATTR_WARN_UNUSED
     Intersection assignRightTurns(const EdgeID via_edge,
                                   Intersection intersection,
                                   const std::size_t up_to) const;