+++ include/extractor/profile_properties.hpp
@@ -12,7 +12,7 @@ struct ProfileProperties
 {
     ProfileProperties()
         : traffic_signal_penalty(0), u_turn_penalty(0), continue_straight_at_waypoint(true),
-          use_turn_restrictions(false)
+          use_turn_restrictions(false), left_hand_driving(false)
     {
     }
 
@@ -36,6 +36,7 @@ struct ProfileProperties
     int u_turn_penalty;
     bool continue_straight_at_waypoint;
     bool use_turn_restrictions;
+    bool left_hand_driving;
 };
 }
 }