+++ include/util/debug.hpp
@@ -5,6 +5,7 @@
 #include "extractor/guidance/turn_lane_data.hpp"
 #include "extractor/query_node.hpp"
 #include "engine/guidance/route_step.hpp"
+#include "util/node_based_graph.hpp"
 #include "util/typedefs.hpp"
 
 #include <iomanip>
@@ -19,11 +20,13 @@ namespace guidance
 {
 inline void print(const engine::guidance::RouteStep &step)
 {
+    const auto lanes = step.intersections.front().lanes;
+
     std::cout << static_cast<int>(step.maneuver.instruction.type) << " "
               << static_cast<int>(step.maneuver.instruction.direction_modifier) << "  "
               << static_cast<int>(step.maneuver.waypoint_type) << " "
-              << " Lanes: (" << static_cast<int>(step.maneuver.lanes.lanes_in_turn) << ", "
-              << static_cast<int>(step.maneuver.lanes.first_lane_from_the_right) << ")"
+              << " Lanes: (" << static_cast<int>(lanes.lanes_in_turn) << ", "
+              << static_cast<int>(lanes.first_lane_from_the_right) << ")"
               << " Duration: " << step.duration << " Distance: " << step.distance
               << " Geometry: " << step.geometry_begin << " " << step.geometry_end
               << " exit: " << step.maneuver.exit << " Intersections: " << step.intersections.size()
@@ -59,6 +62,20 @@ inline void print(const extractor::guidance::Intersection &intersection)
     std::cout << "  Intersection:\n";
     for (const auto &road : intersection)
         std::cout << "\t" << toString(road) << "\n";
+    std::cout << std::flush;
+}
+
+inline void print(const NodeBasedDynamicGraph &node_based_graph,
+                  const extractor::guidance::Intersection &intersection)
+{
+    std::cout << "  Intersection:\n";
+    for (const auto &road : intersection)
+    {
+        std::cout << "\t" << toString(road) << "\n";
+        std::cout << "\t\t"
+                  << node_based_graph.GetEdgeData(road.turn.eid).road_classification.ToString()
+                  << "\n";
+    }
     std::cout << std::flush;
 }