+++ unit_tests/util/static_rtree.cpp
@@ -165,6 +165,9 @@ struct GraphFixture
             // to examine during tests.
             d.forward_segment_id = {pair.second, true};
             d.reverse_segment_id = {pair.first, true};
+            d.fwd_segment_position = 0;
+            d.forward_packed_geometry_id = 0;
+            d.reverse_packed_geometry_id = 0;
             edges.emplace_back(d);
         }
     }